% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePoints.R
\name{correction}
\alias{correction}
\title{Optimizes locations after sample splitting}
\usage{
correction(X, s, t, optimization, batch)
}
\arguments{
\item{X}{data matrix}

\item{s}{first change point candidate in the batch}

\item{t}{second change point candidate in the batch}

\item{optimization}{\code{"L1"} or \code{"L2"}}

\item{batch}{The batch where the candidates came from to reconstruct their
original position in the full data set.}
}
\value{
list containing the candidates in the original data and the corresponding statistic
}
\description{
After sample splitting, we have two batches with candidates each.
The candidates with the lower p-value are chosen. One has to reconstruct the
position of these candidates in the original dataset. Since the actual best
change point could be one point off due to splitting, the neighbors are checked.
}
