% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EvalStatSlow}
\alias{EvalStatSlow}
\title{EvalStatSlow}
\usage{
EvalStatSlow(X, s, t, optimization = "L2")
}
\arguments{
\item{X}{data matrix}

\item{s}{first change point}

\item{t}{second change point}

\item{optimization}{\code{"L1"} or \code{"L2"} loss}
}
\value{
gain/reduction of loss/test statistic for this setup of \code{s} and \code{t}
}
\description{
This function is slow and should be avoided. Evaluates the gain function (test statistic)
\code{optimization = "L2"} will compute the cumulative sums just for this evaluation.
\code{optimization = "L1"} will create a tree just for this evaluation.
}
