% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaxStats}
\alias{MaxStats}
\title{Leading NA}
\usage{
MaxStats(
  X,
  optimization = "L2",
  method = "advanced",
  circular = TRUE,
  min_seg = 2L
)
}
\arguments{
\item{x}{An integer vector}
}
\description{
This function returns a logical vector identifying if
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}
