% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PermTest}
\alias{PermTest}
\title{Maximal CBS statistic}
\usage{
PermTest(
  X,
  boundary,
  cand_stat,
  optimization,
  method,
  alpha,
  nr_perms = 10000L,
  circular = TRUE,
  min_seg = 2L
)
}
\arguments{
\item{X}{data matrix}

\item{boundary}{stopping boundary for early stopping}

\item{cand_stat}{candidate stat, that is, the value from}

\item{optimization}{minimization of \code{L1} or \code{L2} loss}

\item{method}{chose between the fast \code{advanced} or the complete \code{full} search}

\item{alpha}{significance level}

\item{nr_perms}{maximal number of permutations}

\item{circular}{performs binary segmentation if \code{FALSE}. In that case
will return \code{s=0}}

\item{min_seg}{minimal segment length. Note that change points still can be closer together than \code{min_seg}}
}
\value{
list containing the best change point candidates \code{shift}(\eqn{s}) and \code{ind}(\eqn{t}) and the
the corresponding gain \code{stat}
}
\description{
Find the points \code{s,t} so that the loss is reduced most when
splitting the segment in \eqn{X_1,...,X_s,X_{t+1},...,X_T} and \eqn{X_{s+1},...,X_t}.
}
