% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostSelection.R
\name{RemoveSmallJumps}
\alias{RemoveSmallJumps}
\title{Remove Small Jumps}
\usage{
RemoveSmallJumps(X, cps, threshold = 1)
}
\arguments{
\item{X}{data matrix or data vector}

\item{cps}{ordered list of change points the should be reconsidered}

\item{threshold}{jump height relative to standard deviation
so that the change point is accepted}
}
\value{
integer vector with change points with little change removed
}
\description{
Removes all jumps with change in mean less than \code{threshold*SD}
where \code{SD} is a measurement of the standard deviation using the median
absolut deviation estimator.
}
\examples{
set.seed(4)
X <- matrix(rep(c(1,0,1,0), c(80,30,50,40))+rnorm(200,0,0.6), 2, byrow = TRUE)
cps <- optimisticCBS(X, alpha = 0.2)
cps
RemoveSmallJumps(X, cps, 1)
}
