% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing.R
\name{SmoothData}
\alias{SmoothData}
\title{SmoothData}
\usage{
SmoothData(X, R = 3, L = 4, M = 2, times = 1)
}
\arguments{
\item{X}{data matrix or data vector}

\item{R}{radius for smoothing}

\item{L}{threshold for correction}

\item{M}{correction towards median}

\item{times}{number of iterations of the smoothing procedure}
}
\value{
smoothed data in the same form as the input (matrix or vector)
}
\description{
Smoothing technique. See Olshen and Venkatraman "A faster circular binary segmentation algorithm for the analysis of array CGH data".
}
\examples{
set.seed(1)
X <- rt(100,3)
X_smooth <- SmoothData(X)

}
