% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StoppingBoundary}
\alias{StoppingBoundary}
\title{StoppingBoundary}
\usage{
StoppingBoundary(nr_perms = 10000L, alpha = 0.01, eta_star = 5e-04)
}
\arguments{
\item{nr_perms}{maximal number of permutations}

\item{alpha}{significance level}

\item{eta_star}{a control parameter for the probability of early stopping errors}
}
\value{
vector containing \code{nr_perms*alpha} integers
}
\description{
Calculates the early stopping boundary for permutation tests.
Permutation tests stop early if either more than \code{nr_perms*alpha} many
times the permuted data statistic exceeded the original statistic (no change point)
or if the stopping boundary is hit, i.e. the permuted data statistic exceeded
the original statistic not often enough so that is unlikely to be declined (change point).
}
