% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePoints.R
\name{CheckValid}
\alias{CheckValid}
\title{Candidate detection and validation}
\usage{
CheckValid(
  X,
  bd = NULL,
  optimization = "L2",
  method = "advanced",
  select = "mixed",
  alpha = 0.01,
  nr_perms = 10000,
  circular = T,
  min_seg = 2
)
}
\arguments{
\item{X}{data matrix}

\item{bd}{stopping boundary for early stopping}

\item{optimization}{minimization of \code{"L1"} or \code{"L2"} loss}

\item{method}{\code{advanced} or \code{full} search}

\item{select}{Model selection approach \code{perm} (permutation test only), \code{mixed}
(first Wilcoxon test, then permutation test), \code{wilc} (Wilcoxon test only)}

\item{alpha}{Significance level for testing change points}

\item{nr_perms}{Number of permutation for permutation tests}

\item{circular}{\code{circular = T} does CBS, else the test will be for BS}

\item{min_seg}{minimal segment length. Note that change points still can be
closer together than \code{min_seg}}
}
\value{
list containing whether the change points are valid or not and the
corresponding candidates. For (non-circular) BS, there is only one potential
change point, which is stored in \code{t}. In that case \code{s=0.}
}
\description{
Finds the best change point candidates and tests whether they are
statistically significant using Wilcoxon or permutation tests (or a mixture of both).
}
\examples{
set.seed(1)
X <- matrix(rep(c(0,1,0),c(100,20,80)) + rnorm(200,0,0.6), nrow=1)
CheckValid(X, StoppingBoundary())
set.seed(1)
X <- matrix(rnorm(200,0,0.6), nrow=1)
CheckValid(X, StoppingBoundary())
}
