% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePoints.R
\name{optimisticCBS}
\alias{optimisticCBS}
\title{Optimistic Circular Binary Segmentation}
\usage{
optimisticCBS(
  X,
  optimization = "L2",
  method = "advanced",
  select = "mixed",
  alpha = 0.01,
  nr_perms = 10000,
  bd = NULL,
  min_seg = 2
)
}
\arguments{
\item{X}{data vector}

\item{optimization}{minimization of "L1" or "L2" loss}

\item{method}{"advanced" or "full" search}

\item{select}{Model selection approach "perm" (permutation test only), "mixed" (first Wilcoxon test, then permutation test), "wilc" (Wilcoxon test only)}

\item{alpha}{Significance level for testing change points}

\item{nr_perms}{Number of permutation for permutation tests}

\item{bd}{stopping boundary for early stopping}

\item{min_seg}{minimal segment length. Note that change points still can be closer together than \code{min_seg}}
}
\value{
Integer vector containing the change points found in the data
}
\description{
This is the main function of the package.
}
\examples{
set.seed(1)
X <- rep(c(0,1,0),c(100,20,80)) + rnorm(200,0,0.6)
optimisticCBS(X)
}
